Stop looking at environment variables.
Instead, modify Web.config with information from the environment
variables.

This is absolutely necessary for user journey testing.  When we deploy an
app to an IIS instance via Web Deploy, it won't be able to read the
environment variables on our development machine.  Instead, all the
settings must be packed into the Web.config.

I added a bunch of commands to a power shell module to ease the pain of
updating Web.configs.